fix(envd): fix uint64 underflow in heap growth measurement
When the heap shrinks between baseline and measurement, the unsigned
subtraction wraps to a huge value and falsely trips the 50 MiB limit.
Convert to float64 before subtracting so negative growth is handled
correctly.

+++ packages/envd/internal/services/process/start_test.go
@@ -235,7 +235,7 @@ func TestStart_DisconnectStormHeapGrowth(t *testing.T) {
 	var after runtime.MemStats
 	runtime.ReadMemStats(&after)
 
-	heapGrowthMiB := float64(after.HeapInuse-baseline.HeapInuse) / (1024 * 1024)
+	heapGrowthMiB := (float64(after.HeapInuse) - float64(baseline.HeapInuse)) / (1024 * 1024)
 	t.Logf("heap: baseline=%.1f MiB after=%.1f MiB growth=%.1f MiB (%d cycles)",
 		float64(baseline.HeapInuse)/(1024*1024),
 		float64(after.HeapInuse)/(1024*1024),
